fix wrong option datatypes for the scale float filter

+++ src/main/java/io/tiledb/java/api/FloatScalingFilter.java
@@ -26,29 +26,30 @@ public class FloatScalingFilter extends Filter {
    * @param byteWidth The byteWidth param
    * @throws TileDBError
    */
-  public FloatScalingFilter(Context ctx, int offset, int factor, int byteWidth) throws TileDBError {
+  public FloatScalingFilter(Context ctx, double offset, double factor, long byteWidth)
+      throws TileDBError {
     super(ctx, tiledb_filter_type_t.TILEDB_FILTER_SCALE_FLOAT);
     try (NativeArray offsetArray =
             new NativeArray(
                 ctx,
-                new int[] {
+                new double[] {
                   offset,
                 },
-                Integer.class);
+                Double.class);
         NativeArray factorArray =
             new NativeArray(
                 ctx,
-                new int[] {
+                new double[] {
                   factor,
                 },
-                Integer.class);
+                Double.class);
         NativeArray byteWidthArray =
             new NativeArray(
                 ctx,
-                new int[] {
+                new long[] {
                   byteWidth,
                 },
-                Integer.class); ) {
+                Long.class); ) {
       ctx.handleError(
           tiledb.tiledb_filter_set_option(
               ctx.getCtxp(),
@@ -80,14 +81,14 @@ public class FloatScalingFilter extends Filter {
    * @param offset The offset input.
    * @throws TileDBError
    */
-  public void setOffset(Context ctx, int offset) throws TileDBError {
+  public void setOffset(Context ctx, double offset) throws TileDBError {
     try (NativeArray offsetArray =
         new NativeArray(
             ctx,
-            new int[] {
+            new double[] {
               offset,
             },
-            Integer.class)) {
+            Double.class)) {
       ctx.handleError(
           tiledb.tiledb_filter_set_option(
               ctx.getCtxp(),
@@ -107,14 +108,14 @@ public class FloatScalingFilter extends Filter {
    * @param factor The factor input
    * @throws TileDBError
    */
-  public void setFactor(Context ctx, int factor) throws TileDBError {
+  public void setFactor(Context ctx, double factor) throws TileDBError {
     try (NativeArray offsetArray =
         new NativeArray(
             ctx,
-            new int[] {
+            new double[] {
               factor,
             },
-            Integer.class)) {
+            Double.class)) {
       ctx.handleError(
           tiledb.tiledb_filter_set_option(
               ctx.getCtxp(),
@@ -134,14 +135,14 @@ public class FloatScalingFilter extends Filter {
    * @param byteWidth The byteWidth param
    * @throws TileDBError
    */
-  public void setByteWidth(Context ctx, int byteWidth) throws TileDBError {
+  public void setByteWidth(Context ctx, long byteWidth) throws TileDBError {
     try (NativeArray offsetArray =
         new NativeArray(
             ctx,
-            new int[] {
+            new long[] {
               byteWidth,
             },
-            Integer.class)) {
+            Long.class)) {
       ctx.handleError(
           tiledb.tiledb_filter_set_option(
               ctx.getCtxp(),
@@ -168,17 +169,17 @@ public class FloatScalingFilter extends Filter {
    * @return The ByteWidth param
    * @throws TileDBError
    */
-  public int getByteWidth() throws TileDBError {
+  public long getByteWidth() throws TileDBError {
     Context ctx = getCtx();
-    int window;
-    try (NativeArray byteWidthArray = new NativeArray(ctx, 1, Integer.class)) {
+    long window;
+    try (NativeArray byteWidthArray = new NativeArray(ctx, 1, Long.class)) {
       ctx.handleError(
           tiledb.tiledb_filter_get_option(
               ctx.getCtxp(),
               getFilterp(),
               tiledb_filter_option_t.TILEDB_SCALE_FLOAT_BYTEWIDTH,
               byteWidthArray.toVoidPointer()));
-      window = (int) byteWidthArray.getItem(0);
+      window = (long) byteWidthArray.getItem(0);
     }
     return window;
   }
@@ -187,17 +188,17 @@ public class FloatScalingFilter extends Filter {
    * @return The factor param
    * @throws TileDBError
    */
-  public int getFactor() throws TileDBError {
+  public double getFactor() throws TileDBError {
     Context ctx = getCtx();
-    int window;
-    try (NativeArray factorArray = new NativeArray(ctx, 1, Integer.class)) {
+    double window;
+    try (NativeArray factorArray = new NativeArray(ctx, 1, Double.class)) {
       ctx.handleError(
           tiledb.tiledb_filter_get_option(
               ctx.getCtxp(),
               getFilterp(),
               tiledb_filter_option_t.TILEDB_SCALE_FLOAT_FACTOR,
               factorArray.toVoidPointer()));
-      window = (int) factorArray.getItem(0);
+      window = (double) factorArray.getItem(0);
     }
     return window;
   }
@@ -206,17 +207,17 @@ public class FloatScalingFilter extends Filter {
    * @return The offset param
    * @throws TileDBError
    */
-  public int getOffset() throws TileDBError {
+  public double getOffset() throws TileDBError {
     Context ctx = getCtx();
-    int window;
-    try (NativeArray offsetArray = new NativeArray(ctx, 1, Integer.class)) {
+    double window;
+    try (NativeArray offsetArray = new NativeArray(ctx, 1, Double.class)) {
       ctx.handleError(
           tiledb.tiledb_filter_get_option(
               ctx.getCtxp(),
               getFilterp(),
               tiledb_filter_option_t.TILEDB_SCALE_FLOAT_OFFSET,
               offsetArray.toVoidPointer()));
-      window = (int) offsetArray.getItem(0);
+      window = (double) offsetArray.getItem(0);
     }
     return window;
   }
